requirements/base.txt
# Bleeding edge Django
django==1.8.4

# Configuration
django-environ==0.3.0
django-secure==1.0.1
whitenoise==2.0.3


# Forms
django-braces==1.8.1
django-crispy-forms==1.5.0
django-floppyforms==1.5.2

# Models
django-model-utils==2.3.1

# Images
Pillow==2.9.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.23.0

# For the persistence stores
psycopg2==2.6.1

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.8.0

# Time zones support
pytz==2015.4

# Redis support
django-redis==4.2.0
redis>=2.10.0

djangorestframework


# Your custom requirements go here
